reuse VesselsTable in viirs match
