--- apps/fishing-map/features/map/popups/FishingLayers.module.css
@@ -1,12 +1,0 @@
-.modalContainer {
-  padding-bottom: 3rem;
-}
-
-.vesselsMore {
-  opacity: var(--opacity-secondary);
-  cursor: pointer;
-}
-
-.vesselsMore:disabled {
-  cursor: unset;
-}
